feat: refactor select components and update related styles for improved usability
